% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{krpoltext-package}
\alias{krpoltext}
\alias{krpoltext-package}
\title{krpoltext: Korean Political Text Corpora for Reproducible Research}
\description{
Provides easy access to two large-scale Korean political text corpora:
the South Korean Election Campaign Booklet Corpus (49,678 candidates,
2000-2022) and the South Korean Party Statements Corpus (83,201
statements, 2003-2022).
}
\section{Data Descriptor}{

Lim, T.H. (2025). South Korean Election Campaign Booklet and Party
Statements Corpora. \emph{Scientific Data}, 12, 1030.
\doi{10.1038/s41597-025-05220-4}
}

\section{Data Repository}{

\url{https://osf.io/rct9y/} (DOI: 10.17605/OSF.IO/RCT9Y)
}

\seealso{
Useful links:
\itemize{
  \item \url{https://taehyun-lim.github.io/krpoltext/}
  \item \url{https://github.com/taehyun-lim/krpoltext}
  \item \doi{10.1038/s41597-025-05220-4}
  \item \doi{10.5281/zenodo.18704319}
  \item Report bugs at \url{https://github.com/taehyun-lim/krpoltext/issues}
}

}
\author{
\strong{Maintainer}: Tae Hyun Lim \email{tlim@hamilton.edu} (\href{https://orcid.org/0000-0002-0429-2683}{ORCID})

}
\keyword{internal}
